fix: isNotUndefined, defined, arrayIncludes를 internal로 이동합니다.
