+++ src/_internal/index.ts
@@ -16,3 +16,17 @@ export default function assert(condition: boolean, errorMessage?: string): asser
     throw new Error(errorMessage ?? 'Invalid condition');
   }
 }
+
+export function isNotUndefined<T>(value: T | undefined): value is T {
+  return value !== undefined;
+}
+
+export function defined<T>(value: T | undefined): T {
+  assert(value !== undefined);
+
+  return value as T;
+}
+
+export function arrayIncludes<Type>(array: Type[] | readonly Type[], item: unknown, fromIndex?: number): item is Type {
+  return array.includes(item as Type, fromIndex);
+}
