+++ src/standardizePronunciation/rules/transform16th.spec.ts
@@ -1,5 +1,5 @@
+import { defined } from '../../_internal';
 import { disassembleCompleteHangulCharacter } from '../../disassembleCompleteHangulCharacter';
-import { defined } from '../../utils';
 import { transform16th } from './transform16th';
 
 describe('transform16th', () => {
