+++ src/standardizePronunciation/rules/transform17th.spec.ts
@@ -1,5 +1,5 @@
+import { defined } from '../../_internal';
 import { disassembleCompleteHangulCharacter } from '../../disassembleCompleteHangulCharacter';
-import { defined } from '../../utils';
 import { transform17th } from './transform17th';
 
 describe('transform17th', () => {
