+++ src/standardizePronunciation/rules/transform19th.spec.ts
@@ -1,5 +1,5 @@
+import { defined } from '../../_internal';
 import { disassembleCompleteHangulCharacter } from '../../disassembleCompleteHangulCharacter';
-import { defined } from '../../utils';
 import { transform19th } from './transform19th';
 
 describe('transform19th', () => {
