+++ src/standardizePronunciation/rules/transformHardConversion.ts
@@ -1,4 +1,5 @@
-import { arrayIncludes, hasProperty } from '../../utils';
+import { arrayIncludes } from '../../_internal';
+import { hasProperty } from '../../utils';
 import { 된소리, 된소리_받침, 어간_받침 } from '../standardizePronunciation.constants';
 import { ReturnSyllables, Syllable } from './rules.types';
